+++ sorters/expressionsorter.cpp
@@ -1,0 +1,145 @@
+#include "expressionsorter.h"
+#include "qqmlsortfilterproxymodel.h"
+#include <QtQml>
+
+namespace qqsfpm {
+
+/*!
+    \qmltype ExpressionSorter
+    \inherits Sorter
+    \inqmlmodule SortFilterProxyModel
+    \brief Sorts row with a custom sorting
+
+    An ExpressionSorter is a \l Sorter allowing to implement custom sorting based on a javascript expression.
+*/
+
+/*!
+    \qmlproperty expression ExpressionSorter::expression
+
+    An expression to implement custom sorting. It must evaluate to a bool.
+    It has the same syntax has a \l {http://doc.qt.io/qt-5/qtqml-syntax-propertybinding.html} {Property Binding}, except that it will be evaluated for each of the source model's rows.
+    Model data is accessible for both rows with the \c modelLeft, and \c modelRight properties:
+
+    \code
+    sorters: ExpressionSorter {
+        expression: {
+            return modelLeft.someRole < modelRight.someRole;
+        }
+    }
+    \endcode
+
+    The \c index of the row is also available through \c modelLeft and \c modelRight.
+
+    The expression should return \c true if the value of the left item is less than the value of the right item, otherwise returns false.
+
+    This expression is reevaluated for a row every time its model data changes.
+    When an external property (not \c index* or in \c model*) the expression depends on changes, the expression is reevaluated for every row of the source model.
+    To capture the properties the expression depends on, the expression is first executed with invalid data and each property access is detected by the QML engine.
+    This means that if a property is not accessed because of a conditional, it won't be captured and the expression won't be reevaluted when this property changes.
+
+    A workaround to this problem is to access all the properties the expressions depends unconditionally at the beggining of the expression.
+*/
+const QQmlScriptString& ExpressionSorter::expression() const
+{
+    return m_scriptString;
+}
+
+void ExpressionSorter::setExpression(const QQmlScriptString& scriptString)
+{
+    if (m_scriptString == scriptString)
+        return;
+
+    m_scriptString = scriptString;
+    updateExpression();
+
+    Q_EMIT expressionChanged();
+    invalidate();
+}
+
+void ExpressionSorter::proxyModelCompleted(const QQmlSortFilterProxyModel& proxyModel)
+{
+    updateContext(proxyModel);
+}
+
+bool evaluateBoolExpression(QQmlExpression& expression)
+{
+    QVariant variantResult = expression.evaluate();
+    if (expression.hasError()) {
+        qWarning() << expression.error();
+        return false;
+    }
+    if (variantResult.canConvert<bool>()) {
+        return variantResult.toBool();
+    } else {
+        qWarning("%s:%i:%i : Can't convert result to bool",
+                 expression.sourceFile().toUtf8().data(),
+                 expression.lineNumber(),
+                 expression.columnNumber());
+        return false;
+    }
+}
+
+int ExpressionSorter::compare(const QModelIndex& sourceLeft, const QModelIndex& sourceRight, const QQmlSortFilterProxyModel& proxyModel) const
+{
+    if (!m_scriptString.isEmpty()) {
+        QVariantMap modelLeftMap, modelRightMap;
+        QHash<int, QByteArray> roles = proxyModel.roleNames();
+
+        QQmlContext context(qmlContext(this));
+
+        for (auto it = roles.cbegin(); it != roles.cend(); ++it) {
+            modelLeftMap.insert(it.value(), proxyModel.sourceData(sourceLeft, it.key()));
+            modelRightMap.insert(it.value(), proxyModel.sourceData(sourceRight, it.key()));
+        }
+        modelLeftMap.insert("index", sourceLeft.row());
+        modelRightMap.insert("index", sourceRight.row());
+
+        QQmlExpression expression(m_scriptString, &context);
+
+        context.setContextProperty("modelLeft", modelLeftMap);
+        context.setContextProperty("modelRight", modelRightMap);
+        if (evaluateBoolExpression(expression))
+                return -1;
+
+        context.setContextProperty("modelLeft", modelRightMap);
+        context.setContextProperty("modelRight", modelLeftMap);
+        if (evaluateBoolExpression(expression))
+                return 1;
+    }
+    return 0;
+}
+
+void ExpressionSorter::updateContext(const QQmlSortFilterProxyModel& proxyModel)
+{
+    delete m_context;
+    m_context = new QQmlContext(qmlContext(this), this);
+
+    QVariantMap modelLeftMap, modelRightMap;
+    // what about roles changes ?
+
+    for (const QByteArray& roleName : proxyModel.roleNames().values()) {
+        modelLeftMap.insert(roleName, QVariant());
+        modelRightMap.insert(roleName, QVariant());
+    }
+    modelLeftMap.insert("index", -1);
+    modelRightMap.insert("index", -1);
+
+    m_context->setContextProperty("modelLeft", modelLeftMap);
+    m_context->setContextProperty("modelRight", modelRightMap);
+
+    updateExpression();
+}
+
+void ExpressionSorter::updateExpression()
+{
+    if (!m_context)
+        return;
+
+    delete m_expression;
+    m_expression = new QQmlExpression(m_scriptString, m_context, 0, this);
+    connect(m_expression, &QQmlExpression::valueChanged, this, &ExpressionSorter::invalidate);
+    m_expression->setNotifyOnValueChanged(true);
+    m_expression->evaluate();
+}
+
+}
